requirements.txt
autopep8==1.5
bcrypt==3.1.7
bidict==0.19.0
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
paramiko==2.7.1
pycodestyle==2.5.0
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.3
six==1.13.0
Werkzeug==0.16.0
